\name{fit4NM}
\alias{fit4NM}
\title{
Platform for NONMEM
}
\description{
fit4NM is the software for the exploratory analysis for pharmacometrics. 
It helps NONMEM users check their data, fit their model using NONMEM, 
and diagnosis their fitted model. This software is build on R and provides
 a user-friendly graphical interface for the graphics and statistical analysis 
 including running NONMEM.  Therefore, na?ve users can use fit4NM to check their 
 data, fit their model using NONMEM, and diagnosis their model by clicking mouse button.
}

\author{Eun-Kyung Lee,Gyejeong Noh, Hyeong-Seok Lim}

\keyword{ utilities }

\examples{
fit4NM()
}
