% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit4NM.R
\name{fit4NM}
\alias{fit4NM}
\title{Platform for PKPD modelling using NONMEM}
\usage{
fit4NM()
}
\description{
fit4NM is the software for the exploratory analysis for pharmacometrics.
It helps NONMEM users check their data, fit their model using NONMEM,
and diagnosis their fitted model. This software is build on R and provide
a user-friendly graphical interface for the graphics and statistical analysis
including running NONMEM.  Therefore, na?ve users can use fit4NM to check
their data, fit their model using NONMEM, and diagnosis their model by
clicking mouse button.
}
\examples{
fit4NM()
}
\references{
Lee, E., Noh, GJ(2011)
fit4NM: A tool for NONMEM users, PAGANZ 2011, Auckland
}
\keyword{GUI}

